basic game class
