+++ clientV2/index.html
@@ -16,7 +16,7 @@
 			<button id="join-lobby-button">JOIN LOBBY</button>
 		</form>
 	</div>
-	<script type="module" src="src/index.ts"></script>
+	<script type="module" src="src/main.ts"></script>
 </body>
 
 </html>
