--- clientV2/src/bundle.js.LICENSE.txt
@@ -1,5 +1,0 @@
-/**
- * @license
- * Copyright 2010-2025 Three.js Authors
- * SPDX-License-Identifier: MIT
- */
